LayoutTests/imported/w3c/web-platform-tests/trusted-types/empty-default-policy-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'require-trusted-types-for'.

CONSOLE MESSAGE: ReferenceError: Can't find variable: trustedTypes

Harness Error (FAIL), message = ReferenceError: Can't find variable: trustedTypes

TIMEOUT Count SecurityPolicyViolation events. Test timed out

Harness Error (FAIL), message = ReferenceError: Can't find variable: trustedTypes

TIMEOUT Count SecurityPolicyViolation events. Test timed out
